refactored ScalarType
Renamed ScalarType to ScalarDataType and moved it out of the ImageIO object. It is now also private in Scalismo, as it should never be used from outside Scalismo, and hence also not be visible.
