+++ src/main/scala/scalismo/io/ImageIO.scala
@@ -79,105 +79,6 @@ import scala.util.{Failure, Success, Try}
  */
 object ImageIO {
 
-  /**
-   * An enumeration comprising all the data types that we can read and write, in VTK and Nifti formats.
-   */
-  object ScalarType extends Enumeration {
-
-    import NiftiHeader._
-    import VtkHelpers._
-
-    import scala.language.implicitConversions
-
-    protected case class Val[O: Scalar: ClassTag](vtkId: Int, niftiId: Short) extends super.Val
-
-    implicit def valueToVal[T](x: Value): Val[T] = x.asInstanceOf[Val[T]]
-
-    val Byte = Val[Byte](VTK_CHAR, NIFTI_TYPE_INT8)
-    val Short = Val[Short](VTK_SHORT, NIFTI_TYPE_INT16)
-    val Int = Val[Int](VTK_INT, NIFTI_TYPE_INT32)
-    val Float = Val[Float](VTK_FLOAT, NIFTI_TYPE_FLOAT32)
-    val Double = Val[Double](VTK_DOUBLE, NIFTI_TYPE_FLOAT64)
-    val UByte = Val[UByte](VTK_UNSIGNED_CHAR, NIFTI_TYPE_UINT8)
-    val UShort = Val[UShort](VTK_UNSIGNED_SHORT, NIFTI_TYPE_UINT16)
-    val UInt = Val[UInt](VTK_UNSIGNED_INT, NIFTI_TYPE_UINT32)
-
-    /**
-     * Return the ScalarType value corresponding to a given type
-     * @tparam T a scalar type
-     * @return the corresponding ScalarType value
-     * @throws IllegalArgumentException if no corresponding value was found.
-     */
-    def fromType[T: Scalar]: Value = {
-      Scalar[T].scalarType match {
-        case Scalar.ByteScalar   => Byte
-        case Scalar.ShortScalar  => Short
-        case Scalar.IntScalar    => Int
-        case Scalar.FloatScalar  => Float
-        case Scalar.DoubleScalar => Double
-        case Scalar.UByteScalar  => UByte
-        case Scalar.UShortScalar => UShort
-        case Scalar.UIntScalar   => UInt
-        case _                   => throw new IllegalArgumentException(s"Unsupported datatype ${Scalar[T].scalarType}")
-      }
-    }
-
-    /**
-     * Return the ScalarType value corresponding to a given VTK type constant
-     * @param vtkId a VTK type constant
-     * @return the corresponding ScalarType value
-     * @throws IllegalArgumentException if no corresponding value was found.
-     */
-    def fromVtkId(vtkId: Int): Value = {
-      // there are two ways in VTK to represent a (signed) byte.
-      if (vtkId == VTK_SIGNED_CHAR) Byte
-      else
-        values.find(v => v.vtkId == vtkId).getOrElse(throw new IllegalArgumentException(s"Unsupported VTK ID $vtkId"))
-    }
-
-    /**
-     * Return the ScalarType value corresponding to a given Nifti type constant
-     * @param niftiId a Nifti type constant
-     * @return the corresponding ScalarType value
-     * @throws IllegalArgumentException if no corresponding value was found.
-     */
-    def fromNiftiId(niftiId: Short): ScalarType.Value = {
-      values
-        .find(v => v.niftiId == niftiId)
-        .getOrElse(throw new IllegalArgumentException(s"Unsupported Nifti ID $niftiId"))
-    }
-
-    /**
-     * Return the ScalarType value corresponding to the data present in a given file. Only .vtk, .nii and .nia files are supported.
-     * @param file the file to check
-     * @return the scalar type present in the given file, wrapped in a [[scala.util.Success]], or a [[scala.util.Failure]] explaining the error.
-     */
-    def ofFile(file: File): Try[ScalarType.Value] = {
-      val fn = file.getName
-      if (fn.endsWith(".nii") || fn.endsWith(".nia")) {
-        FastReadOnlyNiftiVolume.getScalarType(file).map(ScalarType.fromNiftiId)
-      } else if (fn.endsWith(".vtk")) Try {
-        val reader = new vtkStructuredPointsReader
-        reader.SetFileName(file.getAbsolutePath)
-        reader.Update()
-        val errCode = reader.GetErrorCode()
-        if (errCode != 0) {
-          reader.Delete()
-          throw new IOException(
-            s"Failed to read vtk file ${file.getAbsolutePath}. (error code from vtkReader = $errCode)"
-          )
-        }
-        val st = reader.GetOutput().GetScalarType()
-        reader.Delete()
-        // prevent memory leaks
-        vtkObjectBase.JAVA_OBJECT_MANAGER.gc(false)
-        ScalarType.fromVtkId(st)
-      } else {
-        Failure(new Exception(s"File $file: unsupported file extension"))
-      }
-    }
-  }
-
   trait WriteNifti[D] {
     def write[A: Scalar: ClassTag](img: DiscreteImage[D, A], f: File): Try[Unit]
   }
@@ -251,22 +152,22 @@ object ImageIO {
     }
 
     val result = (for {
-      fileScalarType <- ScalarType.ofFile(file)
+      fileScalarType <- ScalarDataType.ofFile(file)
     } yield {
-      val expectedScalarType = ScalarType.fromType[S]
+      val expectedScalarType = ScalarDataType.fromType[S]
       if (expectedScalarType == fileScalarType) {
         loadAs[S]
       } else {
         val s = implicitly[Scalar[S]]
         fileScalarType match {
-          case ScalarType.Byte   => loadAs[Byte].map(_.map(s.fromByte))
-          case ScalarType.Short  => loadAs[Short].map(_.map(s.fromShort))
-          case ScalarType.Int    => loadAs[Int].map(_.map(s.fromInt))
-          case ScalarType.Float  => loadAs[Float].map(_.map(s.fromFloat))
-          case ScalarType.Double => loadAs[Double].map(_.map(s.fromDouble))
-          case ScalarType.UByte  => loadAs[UByte].map(_.map(u => s.fromShort(u.toShort)))
-          case ScalarType.UShort => loadAs[UShort].map(_.map(u => s.fromInt(u.toInt)))
-          case ScalarType.UInt   => loadAs[UInt].map(_.map(u => s.fromLong(u.toLong)))
+          case ScalarDataType.Byte   => loadAs[Byte].map(_.map(s.fromByte))
+          case ScalarDataType.Short  => loadAs[Short].map(_.map(s.fromShort))
+          case ScalarDataType.Int    => loadAs[Int].map(_.map(s.fromInt))
+          case ScalarDataType.Float  => loadAs[Float].map(_.map(s.fromFloat))
+          case ScalarDataType.Double => loadAs[Double].map(_.map(s.fromDouble))
+          case ScalarDataType.UByte  => loadAs[UByte].map(_.map(u => s.fromShort(u.toShort)))
+          case ScalarDataType.UShort => loadAs[UShort].map(_.map(u => s.fromInt(u.toInt)))
+          case ScalarDataType.UInt   => loadAs[UInt].map(_.map(u => s.fromLong(u.toLong)))
 
           case _ => Failure(new IllegalArgumentException(s"unknown scalar type $fileScalarType"))
         }
@@ -325,22 +226,22 @@ object ImageIO {
     }
 
     val result = (for {
-      fileScalarType <- ScalarType.ofFile(file)
+      fileScalarType <- ScalarDataType.ofFile(file)
     } yield {
-      val expectedScalarType = ScalarType.fromType[S]
+      val expectedScalarType = ScalarDataType.fromType[S]
       if (expectedScalarType == fileScalarType) {
         loadAs[S]
       } else {
         val s = implicitly[Scalar[S]]
         fileScalarType match {
-          case ScalarType.Byte   => loadAs[Byte].map(_.map(s.fromByte))
-          case ScalarType.Short  => loadAs[Short].map(_.map(s.fromShort))
-          case ScalarType.Int    => loadAs[Int].map(_.map(s.fromInt))
-          case ScalarType.Float  => loadAs[Float].map(_.map(s.fromFloat))
-          case ScalarType.Double => loadAs[Double].map(_.map(s.fromDouble))
-          case ScalarType.UByte  => loadAs[UByte].map(_.map(u => s.fromShort(u.toShort)))
-          case ScalarType.UShort => loadAs[UShort].map(_.map(u => s.fromInt(u.toInt)))
-          case ScalarType.UInt   => loadAs[UInt].map(_.map(u => s.fromLong(u.toLong)))
+          case ScalarDataType.Byte   => loadAs[Byte].map(_.map(s.fromByte))
+          case ScalarDataType.Short  => loadAs[Short].map(_.map(s.fromShort))
+          case ScalarDataType.Int    => loadAs[Int].map(_.map(s.fromInt))
+          case ScalarDataType.Float  => loadAs[Float].map(_.map(s.fromFloat))
+          case ScalarDataType.Double => loadAs[Double].map(_.map(s.fromDouble))
+          case ScalarDataType.UByte  => loadAs[UByte].map(_.map(u => s.fromShort(u.toShort)))
+          case ScalarDataType.UShort => loadAs[UShort].map(_.map(u => s.fromInt(u.toInt)))
+          case ScalarDataType.UInt   => loadAs[UInt].map(_.map(u => s.fromLong(u.toLong)))
 
           case _ => Failure(new IllegalArgumentException(s"unknown scalar type $fileScalarType"))
         }
@@ -358,8 +259,8 @@ object ImageIO {
       volume <- FastReadOnlyNiftiVolume.read(file.getAbsolutePath)
       pair <- computeNiftiWorldToVoxelTransforms(volume, favourQform)
     } yield {
-      val expectedScalarType = ScalarType.fromType[S]
-      val foundScalarType = ScalarType.fromNiftiId(volume.header.datatype)
+      val expectedScalarType = ScalarDataType.fromType[S]
+      val foundScalarType = ScalarDataType.fromNiftiId(volume.header.datatype)
       if (expectedScalarType != foundScalarType) {
         throw new IllegalArgumentException(
           s"Invalid scalar type (expected $expectedScalarType, found $foundScalarType)"
@@ -559,7 +460,7 @@ object ImageIO {
       M(3, 3) = 1
 
       // the header
-      volume.header.setDatatype(ScalarType.fromType[S].niftiId)
+      volume.header.setDatatype(ScalarDataType.fromType[S].niftiId)
       volume.header.qform_code = 0
       volume.header.sform_code = 2 // TODO check me that this is right
